laptop-service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2>=2.9
django-cors-headers>=4.3
gunicorn>=21.2
